Initial refactor of frontend into a separate module, probably still improvements to make
